backend/requirements.txt
fastapi==0.115.6
fastapi-cors==0.0.6
uvicorn==0.34.0
pyserial==3.5
